O(n) -> O(k)
For this hunk:
```swift
if i - maxIdx.first! + 1 == k {
      maxIdx.removeFirst()
}
```
Same as: 

```swift
if maxIdx.last! - maxIdx.first! + 1 == k {
      maxIdx.removeFirst()
}
```
As `maxIdx.last! - maxIdx.first! + 1` is always less than or equal to `maxIdx.count`, and will pop the last element when it reaches`k`.
Such that :  `maxIdx.count` is always less than or equal to k.

=> The space complexity is `O(k)`, rather than `O(n)`.